implement time taken feature
It’s now tracked how long users need to answer a question. The average
time taken is shown in the question details, along with other data like
how often this question has been answered correctly. Hopefully this
will make it easier to gauge the difficulty of questions more easily
and find questions that are not 'short' enough.

The average answer time is intentionally not visible to the user to
avoid competition on slow/fast answering.